Quieten default output
Motivation:

The code contains large number of `println` statements.  While this
provides a wealth of information, under some deployments this may
provide too much information and, without a logging framework, tuning
what information to display impossible.

Modification:

Introduce logging via slf4j.  This provides an abstract mechanism to
route logging information to some back-end that decides whether the
message is to be logged.  Logback is introduced as a core dependency
but deployments may override this and introduce a logging tunnel
to integrate logging with their chosen solution.

Result:

The output is quieter by default and may be tunned.  The code also
now supports better integration with existing logging solutions.