Add ability to configure CDMI servlet
Motivation:

The reference implementation provides no mechanism with which
the admin can configure its behaviour.  In particular, the
base directory is hard-coded as '/data'

Modification:

Add the cdmi.properties file that contains configuration options
which are available as options through spring dependency injection.

Result:

The target directory may be configured by the admin.