rewrite functional tests
Motivation:

The current functional tests only check the status code of the HTTP
response and provide a visual output of the headers and any JSON body.
The tests are also written with too much low-level detail and
boiler-plate code, making it hard to see what is being tested.
Together, this limits the ability to use the tests to catch regressions.

Modification:

Add a rudimentary CDMI client to hide details of the requests; this
client supports a fluent interface to aid readability.  The utility
methods for building the server-side harness are moved into their own
class.

Added a set of Matchers that aim to make assertions both easy to read
and easy to write.  These matchers support assertions that check the
HTTP response status code, HTTP header values, and various aspects of
the response entity when parsed as a JSON object.

Rewrote existing functional tests, separating them out into different
classes.  These tests now use the new client, along with an initial
set of assertions on the server responses.

Note that the assertions are based on what the server currently returns.
Several values are suspect and future patches may revise the code,
updating the corresponding assertion(s).

Result:

The code now has an extendible set of functional tests.